[HOTFIX] Fix notebook skipping logic
The existing notebook skipping logic wasn't working correctly for some notebooks (i.e. `sssp_benchmark.ipynb` since it appears at the end of a line with no proceeding space). This PR updates the notebook skipping logic and makes the `SKIPNBS` variable an array for better readability.

The logic was borrowed from the following Stack Overflow answer: https://stackoverflow.com/questions/3685970/check-if-a-bash-array-contains-a-value